+++ src/Components/DetailPage/Modal.tsx
@@ -11,70 +11,21 @@ import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
 import { addLike, getLikes } from "@/util/supabase/detail_supabase_DB";
 import { onClickCopyClipBoardHandler } from "@/util/urlCopy";
 import Loading from "../Loading";
+import useLiked from "@/hooks/community/useLiked";
 
 const Modal = ({ isVisible, onClose, children }: any) => {
+    const { checkLike, id, pending, handleLikeBtn } = useLiked();
+
     if (!isVisible) return null;
     const handleClose = (e: any) => {
         if (e.target.id === "wrapper") {
             onClose();
         }
     };
-    const { cardId: id } = useCardIdStore();
-    const { user } = useUserStore(); //로그인여부 확인
-    const nowUser = user?.user_metadata.email; //로그인 유저의 email
 
-    const queryClient = useQueryClient();
-
-    const { data: likes, isPending } = useQuery({
-        queryKey: [QUERY_KEY.portfolioLikes],
-        queryFn: () => getLikes({ id: "id", value: id }),
-    });
-
-    //좋아요 추가
-    const addLikeMutate = useMutation({
-        mutationFn: addLike,
-        onSuccess: () => {
-            queryClient.invalidateQueries({ queryKey: [QUERY_KEY.portfolioLikes] });
-        },
-    });
-
-    if (isPending) {
+    if (pending) {
         return <Loading />;
     }
-    //좋아요 눌렀는지 확인
-    const checkLike = likes!.find((item) => item === user?.user_metadata.email);
-
-    //좋아요 버튼
-    const handleLikeBtn = () => {
-        if (!user) {
-            return alert("로그인이 필요한 서비스 입니다."); //비회원일시
-        }
-        if (checkLike) {
-            //좋아요를 이미 눌렀을 경우
-            //현재 likes배열에서 내 이메일을 제거
-            const deleteUserEmail = likes!.filter((item) => item !== nowUser);
-            const likeUser = {
-                id,
-                user_email: deleteUserEmail,
-            };
-            addLikeMutate.mutate(likeUser);
-            return;
-        }
-        const nowUserEmail = [...likes!, user!.user_metadata.email]; //현재 유저 email
-        const likeUser = {
-            id,
-            user_email: nowUserEmail,
-        };
-        addLikeMutate.mutate(likeUser);
-    };
-
-    //댓글 버튼
-    const handleCommentBtn = () => {};
-
-    //공유하기 버튼
-    const handleShareBtn = () => {
-        onClickCopyClipBoardHandler(`${process.env.NEXT_PUBLIC_BASE_URL}/share/${id}`);
-    };
 
     return (
         <div
@@ -96,7 +47,7 @@ const Modal = ({ isVisible, onClose, children }: any) => {
                 <div className="flex flex-col items-center justify-center gap-2 ">
                     <button
                         className="bg-white w-[48px] h-[48px] rounded-[999px] p-2 flex items-center justify-center"
-                        onClick={handleLikeBtn}
+                        onClick={() => handleLikeBtn()}
                     >
                         {checkLike ? (
                             <Image src="redHeart.svg" alt="아이콘" width={30} height={30} />
@@ -120,7 +71,7 @@ const Modal = ({ isVisible, onClose, children }: any) => {
                 <div className="flex flex-col items-center justify-center gap-2">
                     <button
                         className="bg-white w-[48px] h-[48px] rounded-[999px] p-2 flex items-center justify-center"
-                        onClick={handleShareBtn}
+                        onClick={() => onClickCopyClipBoardHandler(`${process.env.NEXT_PUBLIC_BASE_URL}/share/${id}`)}
                     >
                         <Image src="share.svg" alt="아이콘" width={25} height={25} />
                     </button>
